requirements.txt
requests
flask==2.1.3
flask-restplus
Werkzeug==2.0.3
python-dotenv
jinja2==3.0.3
itsdangerous==2.1.2
flask-sqlalchemy==2.5.1